requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.23.2
SQLAlchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
email-validator==2.1.0.post1
pydantic[email]==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
qrcode==7.4.2
Pillow==10.1.0
aiosmtplib==2.0.2
fpdf2==2.7.5
orjson==3.9.10
psycopg[binary]==3.1.12
